requirements.txt
amqp==5.0.6
asgiref==3.4.1
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Django==3.2.8
django-filter==21.1
djangorestframework==3.12.4
joblib==1.1.0
kombu==5.1.0
Markdown==3.3.4
nltk==3.6.3
numpy==1.21.2
pandas==1.3.3
prompt-toolkit==3.0.20
psycopg2==2.9.1
python-dateutil==2.8.2
pytz==2021.3
redis==3.5.3
regex==2021.10.8
six==1.16.0
sqlparse==0.4.2
tqdm==4.62.3
vine==5.0.0
wcwidth==0.2.5
PyJWT==2.1.0
djangorestframework-simplejwt==4.8.0